+++ frontend/src/widgets/inputs/TextInputWidget/hooks/index.ts
@@ -5,3 +5,5 @@ export {
   parseShortcut,
   formatShortcutForDisplay,
 } from './useTextInput';
+export { useSyncServerValue } from './useSyncServerValue';
+export { useShortcutKey } from './useShortcutKey';
