+++ frontend/src/widgets/inputs/TextInputWidget/hooks/useTextInput.ts
@@ -1,75 +1,6 @@
 import React, { useRef, useEffect, useCallback } from 'react';
 
-// Utility to detect Mac platform
-const isMac =
-  typeof navigator !== 'undefined' &&
-  /Mac|iPod|iPhone|iPad/.test(navigator.userAgent);
-
-interface ParsedShortcut {
-  ctrl: boolean;
-  shift: boolean;
-  alt: boolean;
-  meta: boolean;
-  key: string;
-}
-
-// Parse the shortcut string into components
-export const parseShortcut = (shortcutStr?: string): ParsedShortcut | null => {
-  if (!shortcutStr) return null;
-  const parts = shortcutStr.toLowerCase().split('+');
-  return {
-    ctrl: !isMac && parts.includes('ctrl'),
-    shift: parts.includes('shift'),
-    alt: parts.includes('alt'),
-    meta: isMac
-      ? parts.includes('ctrl') ||
-        parts.includes('meta') ||
-        parts.includes('cmd') ||
-        parts.includes('command')
-      : false,
-    key: parts[parts.length - 1],
-  };
-};
-
-// Format the shortcut for display as React nodes
-export const formatShortcutForDisplay = (
-  shortcutStr?: string
-): React.ReactNode[] => {
-  if (!shortcutStr) return [];
-  const parts = shortcutStr.split('+').map(p => p.trim());
-  const result: React.ReactNode[] = [];
-
-  parts.forEach((part, index) => {
-    if (index > 0) {
-      result.push('+');
-    }
-
-    if (
-      isMac &&
-      (part.toLowerCase() === 'ctrl' ||
-        part.toLowerCase() === 'cmd' ||
-        part.toLowerCase() === 'command' ||
-        part.toLowerCase() === 'meta')
-    ) {
-      result.push(
-        React.createElement(
-          'span',
-          {
-            key: `meta-${index}`,
-            className: 'inline-flex items-center justify-center',
-          },
-          '⌘'
-        )
-      );
-    } else if (!isMac && part.toLowerCase() === 'ctrl') {
-      result.push('Ctrl');
-    } else {
-      result.push(part.charAt(0).toUpperCase() + part.slice(1));
-    }
-  });
-
-  return result;
-};
+export { parseShortcut, formatShortcutForDisplay } from '../utils/shortcut';
 
 export const useCursorPosition = (
   value?: string,
