+++ frontend/src/widgets/inputs/TextInputWidget/types.ts
@@ -1,6 +1,14 @@
-import React from 'react';
 import { Scales } from '@/types/scale';
-import Icon from '@/components/Icon';
+
+export enum TextInputVariant {
+  Text = 'Text',
+  Textarea = 'Textarea',
+  Email = 'Email',
+  Tel = 'Tel',
+  Url = 'Url',
+  Password = 'Password',
+  Search = 'Search',
+}
 
 export type PrefixSuffix =
   | { type: 'text'; value: string }
@@ -10,14 +18,7 @@ export interface TextInputWidgetProps {
   id: string;
   placeholder?: string;
   value?: string;
-  variant:
-    | 'Text'
-    | 'Textarea'
-    | 'Email'
-    | 'Tel'
-    | 'Url'
-    | 'Password'
-    | 'Search';
+  variant: TextInputVariant;
   disabled: boolean;
   invalid?: string;
   events: string[];
@@ -30,26 +31,3 @@ export interface TextInputWidgetProps {
   maxLength?: number;
   'data-testid'?: string;
 }
-
-/**
- * Renders either text or icon for prefix/suffix display.
- * Uses discriminated union type to ensure only one type can be set.
- */
-export const renderPrefixSuffix = (
-  prefixSuffix?: PrefixSuffix
-): React.ReactNode => {
-  if (!prefixSuffix) return null;
-
-  if (prefixSuffix.type === 'icon') {
-    return React.createElement(Icon, {
-      name: prefixSuffix.value,
-      className: 'w-4 h-4',
-    });
-  }
-
-  return React.createElement(
-    'span',
-    { className: 'text-sm' },
-    prefixSuffix.value
-  );
-};
